% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfo.R
\name{extract_sce}
\alias{extract_sce}
\title{Extract information from SingleCellExperiment object}
\usage{
extract_sce(sce_obj, info_type, info_slot)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{info_type}{A character indicating the type of information to extract. 
"count" represents the count matrix; "label" represents cell labels; "batch"
represent batch labels.}

\item{info_slot}{Slot(s) in assays that stores the count matrix, or slot in 
colData that saves cluster / batch labels.}
}
\value{
A count matrix if info_type is "count"; A data frame if 
info_type is "label"; A vector of characters if info_type is "batch".
}
\description{
This function extracts count matrix, cell labels, and batch labels from 
SingleCellExperiment object.
}
\examples{
## Simulate count matrix, batch, and cell type labels
ng <- 100; nc <- 100
set.seed(1)
X <- cbind(
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 0.1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE),
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE))
colnames(X) <- paste0('c', seq_len(nc))
rownames(X) <- paste0('g', seq_len(ng))
Batches <- rep(seq_len(2), nc/2)
Method1_k2 <- rep(seq_len(2), c(nc/2,nc/2))
Method1_k3 <- sample(seq_len(3), nc, replace = TRUE)

library(SingleCellExperiment)
sim_sce <- SingleCellExperiment(
  list(count = X),
  colData = data.frame(
    Cell_name = colnames(X),
		 Batch = Batches,
  	 Method1_k2 = Method1_k2,
  	 Method1_k3 = Method1_k3),
	rowData = data.frame(Gene_name = rownames(X)))

extract_sce(sim_sce, "count", "count")
extract_sce(sim_sce, "batch", "Batch")
extract_sce(sim_sce, "label", c("Method1_k2", "Method1_k3"))
}
