% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureGeneSelection.R
\name{FeatureGeneSelection}
\alias{FeatureGeneSelection}
\title{Select feature genes}
\usage{
FeatureGeneSelection(
  gcmat,
  method = "wds",
  nfeature = 500,
  batch_label = NULL,
  zp_threshold = 0.95
)
}
\arguments{
\item{gcmat}{A raw count matrix where each row represents a gene, and each column represents a cell.}

\item{method}{A character indicating the method used to select feature genes. "wds" 
(default) represent the working dispersion score proposed in our study; "vst" is the default method 
for feature gene selection in FindVariableFeatures function of Seurat package.}

\item{nfeature}{An integer indicating the number of features to select.}

\item{batch_label}{A vector indicating the batch labels of the cells. The length of batch_label 
should be the same as the number of columns in gcmat.}

\item{zp_threshold}{A number of zero proportion threshold. The range should be (0,1). Genes with 
zero proportion greater than this value will not be considered for feature selection.}
}
\value{
A vector of indices of selected feature genes corresponding to the row indices of input count matrix.
}
\description{
This function selects a subset of feature genes that are expected to express 
differently across cell types before calculating CDI.
}
\examples{
set.seed(100)
X = cbind(matrix(c(rnbinom(2500, size = 1, mu = 0.1), 
                   rnbinom(2500, size = 1, mu = 0.5)), 
                 nrow = 100, byrow = TRUE),
          matrix(c(rnbinom(2500, size = 1, mu = 1), 
                   rnbinom(2500, size = 1, mu = 0.5)), 
                 nrow = 100, byrow = TRUE))
batches = rep(c(1,2), ncol(X)/2)
FeatureGeneSelection(gcmat = X, 
                     batch_label = batches, 
                     nfeature = 20, 
                     zp_threshold = 0.95)
                     
}
\references{
Stuart and Butler et al. Comprehensive Integration of Single-Cell Data. Cell (2019) [Seurat V3]
}
