% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_batch_matrix_celltype.R
\docType{data}
\name{one_batch_matrix_celltype}
\alias{one_batch_matrix_celltype}
\title{Cell type labels of toy count matrix from one batch}
\format{
A vector of 2,000 characters indicating the cell types of the referenece cells. For each cell type, there are 400 cells with the name "type1-5".
}
\usage{
one_batch_matrix_celltype
}
\description{
This is the true label set corresponds to the cells in the one_batch_matrix.
}
\keyword{datasets}
