% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCDI.R
\name{SizeFactor}
\alias{SizeFactor}
\title{This function calculates the size factor of each cell in the UMI count matrix.}
\usage{
SizeFactor(gcmat)
}
\arguments{
\item{gcmat}{A raw UMI count matrix where each row represents a gene, and 
each column represents a cell. The genes should be those before feature gene 
selection.}
}
\value{
A numeric vector indicating the size factor of the cells. 
This should be one of the inputs of the function CalculateCDI.
}
\description{
Different cells have different library size. 
This function calculates the size factor of each cell in the UMI count matrix 
to capture the variation in cell library size.
}
\examples{

X = cbind(matrix(c(rnbinom(2500, size = 1, mu = 0.1), 
                   rnbinom(2500, size = 1, mu = 0.5)), 
                 nrow = 100, byrow = TRUE),
          matrix(c(rnbinom(2500, size = 1, mu = 1), 
                   rnbinom(2500, size = 1, mu = 0.5)), 
                 nrow = 100, byrow = TRUE))
SizeFactor(X)

}
