% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_batch_matrix_batch}
\alias{two_batch_matrix_batch}
\title{Batch labels of simulated count matrix from two batches}
\format{
A vector of 2,000 characters indicating the batch labels of cells 
in two_batch_matrix. 
For each batch, there are 1,000 cells.
}
\usage{
data(two_batch_matrix_batch)
}
\description{
This is the batch label set corresponds to the cells in the two_batch_matrix. 
The code simulating this vector can be found in data-raw/two_batch_simulation.R.
}
\keyword{datasets}
