% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_batch_matrix_celltype}
\alias{two_batch_matrix_celltype}
\title{Cell type labels of simulated count matrix from two batches}
\format{
A vector of 2,000 characters indicating the cell types of cells 
in two_batch_matrix. 
For each cell type, there are 400 cells with the name "type1-5".
}
\usage{
two_batch_matrix_celltype
}
\description{
This is the true label set corresponds to the cells in the two_batch_matrix.
}
\keyword{datasets}
