% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureGeneSelection.R
\name{feature_gene_selection}
\alias{feature_gene_selection}
\title{Select feature genes}
\usage{
feature_gene_selection(
  gcmat = NULL,
  method = "wds",
  nfeature = 500,
  batch_label = NULL,
  zp_threshold = 0.95
)
}
\arguments{
\item{gcmat}{A raw count matrix where each row represents a gene, 
and each column represents a cell.}

\item{method}{A character indicating the method used to select 
feature genes. "wds" (default) represent the working dispersion score 
proposed in our study; "vst" is the default method.
for feature gene selection in FindVariableFeatures function of Seurat package.}

\item{nfeature}{An integer indicating the number of features to select.}

\item{batch_label}{A vector indicating the batch labels of the cells. 
The length of batch_label 
should be the same as the number of columns in gcmat.}

\item{zp_threshold}{A number of zero proportion threshold. The range 
should be (0,1). Genes with 
zero proportion greater than this value will not be considered 
for feature selection.}
}
\value{
A vector of indices of selected feature genes corresponding to the 
row indices of input count matrix.
}
\description{
This function selects a subset of feature genes that are expected to express 
differently across cell types before calculating CDI.
}
\examples{
## Simulate a matrix of 100 rows (genes), where the first 50 genes have
## different mean expression levels.
## Apply feature_gene_selection to select genes
ng = 100; nc = 100
set.seed(1)
X <- cbind(
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 0.1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE),
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE))
colnames(X) <- paste0('c', seq_len(nc))
rownames(X) <- paste0('g', seq_len(ng))
Batches <- rep(seq_len(2), nc/2)

## Input: matrix
feature_gene_selection(
	gcmat = X,
	batch_label = Batches,
	nfeature = 20,
	zp_threshold = 0.95)


## Input: SingleCellExperiment object
library(SingleCellExperiment)
sim_sce <- SingleCellExperiment(
  list(count = X),
  colData = data.frame(
  	Cell_name = colnames(X),
		batch = Batches),
	rowData = data.frame(gene_name = rownames(X)))
feature_gene_selection(
	gcmat = extract_sce(sim_sce, "count", "count"),
	batch_label = extract_sce(sim_sce, "batch", "batch"),
	nfeature = 20,
	zp_threshold = 0.95)



## Input: Seurat object
library(Seurat)
library(SeuratObject)
sim_seurat <- CreateSeuratObject(counts = as.data.frame(X))
sim_seurat <- AddMetaData(sim_seurat, colnames(X), "Cell_name")
sim_seurat <- AddMetaData(sim_seurat, Batches, "Batch")

feature_gene_selection(
	gcmat = extract_seurat(sim_seurat, "count", "counts"),
	batch_label = extract_seurat(sim_seurat, "batch", "Batch"),
	nfeature = 20,
	zp_threshold = 0.95)
                     
}
\references{
Stuart and Butler et al. Comprehensive Integration of 
Single-Cell Data. Cell (2019) [Seurat V3]
}
