% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_batch_matrix.R
\docType{data}
\name{one_batch_matrix}
\alias{one_batch_matrix}
\title{Toy count matrix from one batch}
\format{
A matrix with 3,000 genes (rows) and 2,000 cells (columns). 
The cells are from 5 cell type, and each cell type contains 400 cells.
}
\usage{
one_batch_matrix
}
\description{
Each gene in this matrix is simulated from a mixture of Negative Binomial distribution.
}
\keyword{datasets}
