% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_batch_matrix_celltype}
\alias{one_batch_matrix_celltype}
\title{Cell type labels of simulated count matrix from one batch}
\format{
A vector of 2,000 characters indicating the cell types of the 
reference cells. For each cell type, there are 400 cells with the name 
"type1-5".
}
\usage{
data(one_batch_matrix_celltype)
}
\description{
This is the true label set corresponds to the cells in the one_batch_matrix. 
The code simulating this vector can be found in data-raw/one_batch_simulation.R.
}
\keyword{datasets}
