% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCDI.R
\name{size_factor}
\alias{size_factor}
\title{Size factor of each cell}
\usage{
size_factor(gcmat)
}
\arguments{
\item{gcmat}{A raw UMI count matrix where each row represents a gene, and 
each column represents a cell. The genes should be those before feature gene 
selection.}
}
\value{
A numeric vector indicating the size factor of the cells. 
This should be one of the inputs of the function calculate_CDI.
}
\description{
Different cells have different library size. 
This function calculates the size factor of each cell in the UMI count matrix 
to capture the variation in cell library size.
}
\examples{

ng = 100; nc = 100
set.seed(1)
X <- cbind(
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 0.1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE),
	matrix(
		c(rnbinom(ng*nc/4, size = 1, mu = 1),
			rnbinom(ng*nc/4, size = 1, mu = 0.5)),
		nrow = ng, 
		byrow = TRUE))
colnames(X) <- paste0('c', seq_len(nc))
rownames(X) <- paste0('g', seq_len(ng))

## Input: matrix
size_factor(X)

## Input: SingleCellExperiment object
library(SingleCellExperiment)
sim_sce <- SingleCellExperiment(
  list(count = X),
  colData = data.frame(Cell_name = colnames(X)),
	 rowData = data.frame(Gene_name = rownames(X)))
size_factor(extract_sce(sim_sce, "count", "count"))

## Input: Seurat object
library(Seurat)
library(SeuratObject)
sim_seurat <- CreateSeuratObject(counts = as.data.frame(X))
sim_seurat <- AddMetaData(sim_seurat, colnames(X), "Cell_name")

}
