% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_batch_matrix.R
\docType{data}
\name{two_batch_matrix}
\alias{two_batch_matrix}
\title{Toy count matrix from two batches}
\format{
A matrix with 3,000 genes (rows) and 2,000 cells (columns). 
The cells are from 5 cell type, and each cell type contains 400 cells. 
The first 1,000 cells are from batch 1, and the second 1,000 cells are from batch 2.
}
\usage{
two_batch_matrix
}
\description{
This matrix is simulated from NB distribution.
}
\keyword{datasets}
