% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_batch_matrix_batch.R
\docType{data}
\name{two_batch_matrix_batch}
\alias{two_batch_matrix_batch}
\title{Batch labels of toy count matrix from two batches}
\format{
A vector of 2,000 characters indicating the batch labels of cells in two_batch_matrix. 
For each batch, there are 1,000 cells.
}
\usage{
two_batch_matrix_batch
}
\description{
This is the batch label set corresponds to the cells in the two_batch_matrix.
}
\keyword{datasets}
